refactor: Mark these functions as test only
